requirements.txt
torch
torchvision
torchaudio
numpy
opencv-python
Pillow
selenium
webdriver-manager
pandas
